pyproject.toml
[tool.poetry]
name = "jqlv2"
version = "0.10"
description = ""
authors = ["Jason Estrada"]
license = "One Network Enterprises"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai-tools = "^0.2.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
crewai = "^0.30.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
